fix: uniform client API (#31)
* fix: refactor to uniform API

* refactor: change package to influxdb3

* refactor: rename client_write file to just write

* refactor: add query and write methods with options

* fix: remove Query method parameter with unclear meaning and no tests
