+++ README.md
@@ -82,7 +82,7 @@ import (
 )
 ```
 
-Create `influx.Client` with `New` function. Make sure to `Close` client after with `defer` keyword.
+Create `influxdb3.Client` with `New` function. Make sure to `Close` client after with `defer` keyword.
 
 ```go
 url := os.Getenv("INFLUXDB_URL")
@@ -90,12 +90,13 @@ token := os.Getenv("INFLUXDB_TOKEN")
 database := os.Getenv("INFLUXDB_DATABASE")
 
 // Create a new client using an InfluxDB server base URL and an authentication token
-client, err := influx.New(influx.Configs{
-    HostURL: url,
-    AuthToken: token,
+client, err := influxdb3.New(influxdb3.ClientConfig{
+    Host: url,
+    Token: token,
+    Database: database,
 })
 // Close client at the end and escalate error if present
-defer func (client *influx.Client)  {
+defer func (client *influxdb3.Client)  {
     err := client.Close()
     if err != nil {
         panic(err)
@@ -107,7 +108,7 @@ The `client` can be now used to insert data using [line-protocol](https://docs.i
 
 ```go
 line := "stat,unit=temperature avg=23.5,max=45.0"
-err = client.Write(context.Background(), database, []byte(line))
+err = client.Write(context.Background(), []byte(line))
 ```
 
 Fetch data using FlightSQL query and print result.
@@ -122,7 +123,7 @@ query := `
         "unit" IN ('temperature')
 `;
 
-iterator, err := client.Query(context.Background(), database, query, nil)
+iterator, err := client.Query(context.Background(), query)
 
 if err != nil {
     panic(err)
