+++ example/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/InfluxCommunity/influxdb3-go/influx"
+	"github.com/InfluxCommunity/influxdb3-go/influxdb3"
 )
 
 func main() {
@@ -16,16 +16,17 @@ func main() {
 	database := os.Getenv("INFLUXDB_DATABASE")
 
 	// Create a new client using an InfluxDB server base URL and an authentication token
-	client, err := influx.New(influx.Configs{
-		HostURL:   url,
-		AuthToken: token,
+	client, err := influxdb3.New(influxdb3.ClientConfig{
+		Host:     url,
+		Token:    token,
+		Database: database,
 	})
 
 	if err != nil {
 		panic(err)
 	}
 	// Close client at the end and escalate error if present
-	defer func(client *influx.Client) {
+	defer func(client *influxdb3.Client) {
 		err := client.Close()
 		if err != nil {
 			panic(err)
@@ -33,23 +34,23 @@ func main() {
 	}(client)
 
 	// Create point using full params constructor
-	p := influx.NewPoint("stat",
+	p := influxdb3.NewPoint("stat",
 		map[string]string{"unit": "temperature"},
 		map[string]interface{}{"avg": 24.5, "max": 45.0},
 		time.Now())
 	// write point synchronously
-	err = client.WritePoints(context.Background(), database, p)
+	err = client.WritePoints(context.Background(), p)
 	if err != nil {
 		panic(err)
 	}
 	// Create point using fluent style
-	p = influx.NewPointWithMeasurement("stat").
+	p = influxdb3.NewPointWithMeasurement("stat").
 		AddTag("unit", "temperature").
 		AddField("avg", 23.2).
 		AddField("max", 45.0).
 		SetTimestamp(time.Now())
 	// write point synchronously
-	err = client.WritePoints(context.Background(), database, p)
+	err = client.WritePoints(context.Background(), p)
 	if err != nil {
 		panic(err)
 	}
@@ -62,13 +63,13 @@ func main() {
 		Time  time.Time `lp:"timestamp"`
 	}{"stat", "temperature", 22.3, 40.3, time.Now()}
 	// Write point
-	err = client.WriteData(context.Background(), database, sensorData)
+	err = client.WriteData(context.Background(), sensorData)
 	if err != nil {
 		panic(err)
 	}
 	// Or write directly line protocol
 	line := fmt.Sprintf("stat,unit=temperature avg=%f,max=%f", 23.5, 45.0)
-	err = client.Write(context.Background(), database, []byte(line))
+	err = client.Write(context.Background(), []byte(line))
 	if err != nil {
 		panic(err)
 	}
@@ -83,7 +84,7 @@ func main() {
     "unit" IN ('temperature')
   `
 
-	iterator, err := client.Query(context.Background(), database, query)
+	iterator, err := client.Query(context.Background(), query)
 
 	if err != nil {
 		panic(err)
